+++ griptape/config/__init__.py
@@ -4,11 +4,12 @@ from .structure_global_drivers_config import StructureGlobalDriversConfig
 from .structure_task_memory_extraction_engine_csv_config import StructureTaskMemoryExtractionEngineCsvConfig
 from .structure_task_memory_extraction_engine_json_config import StructureTaskMemoryExtractionEngineJsonConfig
 from .structure_task_memory_extraction_engine_config import StructureTaskMemoryExtractionEngineConfig
-from .structure_task_memory_config import StructureTaskMemoryConfig
 from .structure_task_memory_query_engine_config import StructureTaskMemoryQueryEngineConfig
 from .structure_task_memory_summary_engine_config import StructureTaskMemorySummaryEngineConfig
+from .structure_task_memory_config import StructureTaskMemoryConfig
 from .base_structure_config import BaseStructureConfig
 
+from .structure_config import StructureConfig
 from .openai_structure_config import OpenAiStructureConfig
 from .amazon_bedrock_structure_config import AmazonBedrockStructureConfig
 
@@ -23,6 +24,7 @@ __all__ = [
     "StructureTaskMemoryExtractionEngineConfig",
     "StructureTaskMemoryExtractionEngineCsvConfig",
     "StructureTaskMemoryExtractionEngineJsonConfig",
+    "StructureConfig",
     "OpenAiStructureConfig",
     "AmazonBedrockStructureConfig",
 ]