+++ griptape/config/base_config.py
@@ -1,5 +1,6 @@
-from attrs import define
 from abc import ABC
+
+from attrs import define
 
 from griptape.mixins.serializable_mixin import SerializableMixin