+++ griptape/config/structure_task_memory_extraction_engine_config.py
@@ -1,12 +1,14 @@
-from attrs import define, field
+from attrs import Factory, define, field
 
+from griptape.config import StructureTaskMemoryExtractionEngineCsvConfig, StructureTaskMemoryExtractionEngineJsonConfig
 from griptape.mixins.serializable_mixin import SerializableMixin
-
-from .structure_task_memory_extraction_engine_csv_config import StructureTaskMemoryExtractionEngineCsvConfig
-from .structure_task_memory_extraction_engine_json_config import StructureTaskMemoryExtractionEngineJsonConfig
 
 
 @define(kw_only=True)
 class StructureTaskMemoryExtractionEngineConfig(SerializableMixin):
-    csv: StructureTaskMemoryExtractionEngineCsvConfig = field(metadata={"serializable": True})
-    json: StructureTaskMemoryExtractionEngineJsonConfig = field(metadata={"serializable": True})
+    csv: StructureTaskMemoryExtractionEngineCsvConfig = field(
+        default=Factory(lambda: StructureTaskMemoryExtractionEngineCsvConfig()), metadata={"serializable": True}
+    )
+    json: StructureTaskMemoryExtractionEngineJsonConfig = field(
+        default=Factory(lambda: StructureTaskMemoryExtractionEngineJsonConfig()), metadata={"serializable": True}
+    )