+++ griptape/config/structure_task_memory_query_engine_config.py
@@ -1,10 +1,19 @@
-from attrs import define, field
+from attrs import Factory, define, field
 
-from griptape.drivers import BaseVectorStoreDriver, BasePromptDriver
+from griptape.drivers import (
+    BasePromptDriver,
+    BaseVectorStoreDriver,
+    NopVectorStoreDriver,
+    NopEmbeddingDriver,
+    NopPromptDriver,
+)
 from griptape.mixins.serializable_mixin import SerializableMixin
 
 
 @define(kw_only=True)
 class StructureTaskMemoryQueryEngineConfig(SerializableMixin):
-    prompt_driver: BasePromptDriver = field(metadata={"serializable": True})
-    vector_store_driver: BaseVectorStoreDriver = field(metadata={"serializable": True})
+    prompt_driver: BasePromptDriver = field(default=Factory(lambda: NopPromptDriver()), metadata={"serializable": True})
+    vector_store_driver: BaseVectorStoreDriver = field(
+        default=Factory(lambda: NopVectorStoreDriver(embedding_driver=NopEmbeddingDriver())),
+        metadata={"serializable": True},
+    )