+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `AzureMongoDbVectorStoreDriver` for using CosmosDB with MongoDB vCore API.
 - `vector_path` field on `MongoDbAtlasVectorStoreDriver`.
 - `LeonardoImageGenerationDriver` supports image to image generation.
+- `OpenAiStructureConfig` for providing Structures with all OpenAi Driver configuration. 
+- `AmazonBedrockStructureConfig` for providing Structures with all Amazon Bedrock Driver configuration. 
+- `StructureConfig` for building your own Structure configuration.
 
 ### Fixed 
 - `BedrockStableDiffusionImageGenerationModelDriver` request parameters for SDXLv1.
@@ -21,6 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **BREAKING**: Make `index_name` on `MongoDbAtlasVectorStoreDriver` a required field.
 - **BREAKING**: Remove `create_index()` from `MarqoVectorStoreDriver`, `OpenSearchVectorStoreDriver`, `PineconeVectorStoreDriver`, `RedisVectorStoreDriver`.
 - **BREAKING**: `ImageLoader().load()` now accepts image bytes instead of a file path.
+- Deprecated `Structure.prompt_driver` in favor of `Structure.global_drivers.prompt_driver`.
+- Deprecated `Structure.embedding_driver` in favor of `Structure.global_drivers.embedding_driver`.
+- Deprecated `Structure.stream` in favor of `Structure.global_drivers.prompt_driver.stream`.
 
 ## [0.22.3] - 2024-01-22